src/training/requirements.txt
# Training Requirements
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azureml-mlflow>=1.53.0
mlflow>=2.9.0
feast[redis]>=0.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
onnx>=1.15.0
onnxruntime>=1.16.0
skl2onnx>=1.16.0
joblib>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data Ingestion
requests>=2.31.0
tqdm>=4.66.0
